Combat flow correction
