+++ lang/en.json
@@ -205,5 +205,7 @@
 "SETTINGS.DefaultDifficulty": "Default check difficulty",
 "SETTINGS.DefaultDifficultyHint": "You can choose the default difficulty for checks. Unknown will roll checks without the player knowing the check difficulty.",
 "SETTINGS.CheckDifficultyRegular": "Default difficulty",
-"SETTINGS.CheckDifficultyUnknown": "Unknown difficulty"
+"SETTINGS.CheckDifficultyUnknown": "Unknown difficulty",
+"SETTINGS.UseToken":  "Use token image",
+"SETTINGS.UseTokenHint": "Use token image instead of portraits in chat-cards"
 }
