+++ module/chat.js
@@ -1,7 +1,9 @@
 import { CoC7Dice } from './dice.js';
 import { CoC7Check } from './check.js';
 import { COC7 } from './config.js';
-import { CoC7MeleeInitiator, CoC7MeleeTarget, CoC7MeleeResoltion } from './chat/meleecombat.js';
+import { CoC7MeleeInitiator} from './chat/combat/melee-initiator.js';
+import {  CoC7MeleeTarget} from './chat/combat/melee-target.js';
+import { CoC7MeleeResoltion } from './chat/combat/melee-resolution.js';
 import { CoC7RangeInitiator } from './chat/rangecombat.js';
 import { CoC7Roll, chatHelper } from './chat/helper.js';
 import { CoC7DamageRoll } from './chat/damagecards.js';
@@ -77,8 +79,10 @@ export class CoC7Chat{
 						}
 						else { 
 							const initiator = CoC7MeleeInitiator.getFromMessageId( chatMessage.id);
-							const resolutionCard = new CoC7MeleeResoltion( chatMessage.id, null, initiator.resolutionCard);
-							resolutionCard.resolve();
+							if( initiator.resolutionCard){
+								const resolutionCard = new CoC7MeleeResoltion( chatMessage.id, null, initiator.resolutionCard);
+								resolutionCard.resolve();
+							}
 						}
 					}
 					if( card.classList.contains('target')){
