+++ module/chat/helper.js
@@ -15,14 +15,10 @@ export class chatHelper{
 
 	static getActorFromKey(key) {
 
+		if( !key) return null;
 		// Case 1 - a synthetic actor from a Token
 		if (key.includes('.')) {
-			const [sceneId, tokenId] = key.split('.');
-			const scene = game.scenes.get(sceneId);
-			if (!scene) return null;
-			const tokenData = scene.getEmbeddedEntity('Token', tokenId);
-			if (!tokenData) return null;
-			const token = new Token(tokenData);
+			const token = chatHelper.getTokenFromKey(key);
 			return token.actor;
 		}
 
@@ -46,6 +42,7 @@ export class chatHelper{
 	}
 
 	static getTokenFromKey( key){
+		if( !key) return null;
 		if (key.includes('.')) {
 			const [sceneId, tokenId] = key.split('.');
 			const scene = game.scenes.get(sceneId);
@@ -55,13 +52,53 @@ export class chatHelper{
 			const token = new Token(tokenData);
 			return token;
 		} else {
-			const scene = game.scenes.active;
-			if (!scene) return null;
-			const tokens = scene.data.tokens.filter( t => key === t.actorId);
-			if( !tokens.length) return null;
-			const token = new Token(tokens[0]);
-			return token;
+			const actor = game.actors.get( key);
+			return chatHelper.getActorToken( actor, false);
 		}
+	}
+
+	static getActorToken( actor, verbose = true){
+		if( !actor) return null;
+		// Case 0 - Actor is a token (synthetic actor), return that token.
+		if(actor.isToken) return actor.token;
+		else{
+			// Case 1 - Actor is not a token, find if a token exist for that actor.
+			const actorTokens = actor.getActiveTokens();
+			if( actorTokens.length){
+				// Case 1.1 - If he has only one Token return it.
+				if( 1 === actorTokens.length) return actorTokens[0];
+
+				// Case 1.2 - Actor has multiple tokens, find if one of them is the controlled token.
+				const controlledTokens = actorTokens.filter( t => t._controlled);
+				if( controlledTokens.length){
+					// Return the 1st controlled token, rise a warning if he has multiple controlled tokens.
+					if( verbose && controlledTokens.length > 1) ui.notifications.warn( `Actor ${actor.name} has ${controlledTokens.length} controlled tokens. Using the first found`);
+					return controlledTokens[0];
+				}
+
+				// Case 1.3 actor doesn't have any active token. Return the first valid token for that actor and raise a warning.
+				if( verbose) ui.notifications.warn( `Actor ${actor.name} doesn't have any controlled token. Using first token found.`);
+				return actorTokens[0];
+			}
+
+			if( verbose) ui.notifications.error( `Could not fin any token for ${actor.name}.`);
+			return null;
+		}
+	}
+
+	static getActorImgFromKey( actorKey){
+		if( !actorKey) return null;
+		if( game.settings.get('CoC7', 'useToken')){
+			// Try to find a token.
+			const token = chatHelper.getTokenFromKey(actorKey);
+			if( token) return token.data.img;
+		}
+		const actor = chatHelper.getActorFromKey(actorKey);
+		if( game.settings.get('CoC7', 'useToken')){
+			// if no token found for that actor return the prototype token image.
+			if( actor.data.token) return actor.data.token.img;
+		}
+		return actor.data.img;
 	}
 
 	static getDistance( startToken, endToken){
